refactor: use htmlreport redux stuff in DSR and RRS sections
