+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-11T17:52:10.238Z\n"
-"PO-Revision-Date: 2019-02-11T17:52:10.238Z\n"
+"POT-Creation-Date: 2019-02-11T18:16:22.102Z\n"
+"PO-Revision-Date: 2019-02-11T18:16:22.102Z\n"
 
 msgid "Cancel"
 msgstr ""
